requirements.txt
python-telegram-bot
langgraph
google-cloud-speech
google-generativeai
sqlalchemy
alembic
jdatetime
pytest
python-dotenv
psutil
pydantic
pydantic-settings
pytz
langgraph-checkpoint-sqlite
langchain-google-genai